New attempt to enable switching between testing and RPA modes
The old approach used a Python-Markdown plugin for conversion and this
one used a MkDocs hook. Both approaches work, but the former required
the plugin to be importable.

MkDocs doesn't seem to have way to add new directories to PYTHONPATH,
and making the plugin installable would have been too much work. Hooks
are loaded without extra configuration, and the MkDocs hoop API is
also simpler than the Python-Markdownn plugin API.