Better approach to handle `*** Test Cases ***` header.
Using Javascript to toggle between `*** Test Cases ***` and `*** Tasks
***`, based on the test/task mode, caused flicker, because the former
header was rendered first and then changed to the latter.

The new approach modifies the generated HTML so that it contains both
headers with `test` and `task` classes that we use with other altering
content as well. This way showing/hiding headers is handled using the
same logic as hiding other such content and the front-end code gets
simpler. More importantly, flickering is gone!